feat(mobile): exposed EstimateFee
